Fix passcode not shown
- add small dismiss animation

Instead of modally displaying the passcode on top of the current VC use the privacy window to display the passcode on top of it.
This allows us to keep whatever VCs are below that intact and active (including system modals like QuickLook markup, Photo select, UIActivityVC, etc.), because the window can be displayed on top of everything (even alerts).
It also improves privacy, because tho the blur effect on top of whatever file is opened looks kinda cool, the purpose of adding a passcode is so that others can't see my files. The blur view gave away more information about what files are there / how many.

Signed-off-by: Henrik Storch <henrik.storch@nextcloud.com>